third_party/blink/web_tests/http/tests/devtools/geolocation-emulation-tests-expected.txt

Tests that geolocation emulation with latitude and longitude works as expected.


Running: testGeolocationUnavailable
Position unavailable

Running: testOverridenGeolocation
Latitude: 50 Longitude: 100

Running: testInvalidParam
error: Request Emulation.setGeolocationOverride failed. {"code":-32602,"message":"Invalid parameters","data":"Failed to deserialize params.latitude - BINDINGS: double value expected at position 17"}

Running: testInvalidGeolocation
Latitude: 50 Longitude: 100

Running: testNoOverride
Override was cleared correctly.
